Extract a shared query-consumption engine (reactions + sources)
Query-consuming sources previously used a bespoke, weaker forwarder: a bare live
subscribe with no outbox/snapshot catch-up, no checkpoint/restart resume, and no
gap recovery — so results emitted before the source subscribed (or after a
restart) were silently missed. Reactions already had a robust consumption engine
for exactly this.

This unifies them into ONE shared engine in `lib/src/consumption/`:

- `ConsumptionEngine` holds the orchestration extracted verbatim from the reaction
  manager (subscribe + per-query fresh-start/outbox catch-up, checkpointing,
  hash-mismatch + gap/lag recovery, the bootstrap gate, and metrics). Both
  `ReactionManager` and `SourceManager` embed one and share the same dependency
  Arcs, so behavior is preserved exactly.
- `QueryConsumer` trait is the interface the engine drives; components adapt via
  thin `ReactionConsumer`/`SourceConsumer` wrappers, leaving the `Reaction` and
  `Source` traits (and their base types) untouched.
- Terminology: the engine uses "catch-up" (`CatchupContext`) for the input side;
  "bootstrap" stays reserved for the source output `BootstrapProvider`.
  `Reaction::bootstrap` is preserved as a thin alias (its `catch_up` delegates to
  it), so the reaction API and FFI are unchanged.

Because the engine runs host-side and only forwards results through the existing
`Source::enqueue_query_result`, dynamically-loaded (FFI) sources gain the same
catch-up/checkpoint/gap-recovery guarantees with no new vtable functions — the
outbox, checkpoints, and gap detection are all host-side.

Validated: full drasi-lib suite (868 tests incl. reaction manager + checkpoint
e2e), the state-machine crate (incl. a new regression test proving a source
catches up on a query result emitted before it subscribed), and the dynamic
cdylib e2e. Clippy + fmt clean.
